Guess the Number Game - Java
This Java exercise prompts the user to guess a number and then checks if the guess is correct. If the guess is incorrect, the program provides feedback to the user and prompts them to guess again until they guess correctly. The program uses a while loop to continue prompting the user until they guess the correct number. This exercise is a good way to practice using basic Java syntax, including input/output and control flow statements like if-else and while loops.