Reading and Printing Names in Reverse Order - Java
This Java program reads in a number from the user via the Scanner class and uses it to determine how many names to read in from the user. It then initializes an ArrayList of strings and adds each name to the list using the add method. Finally, the program prints out the list of names in the order they were entered, and then prints out the list in reverse order using a for loop that iterates through the list in reverse order. The teclado.nextLine() method is used to consume the newline character after reading in the integer value to avoid any issues with reading in the names. This program can be useful for practicing input/output and working with ArrayLists in Java.